Reference

Contents

Reference

Contents

Index

Index

AdsorbedSolutionTheory.DualSiteLangmuir

AdsorbedSolutionTheory.Langmuir

AdsorbedSolutionTheory.Toth

AdsorbedSolutionTheory.Unilan

AdsorbedSolutionTheory.f∂f

AdsorbedSolutionTheory.f∂f∂2f

AdsorbedSolutionTheory.henry_coefficient

AdsorbedSolutionTheory.isosteric_heat

AdsorbedSolutionTheory.loading

AdsorbedSolutionTheory.nlsolve

AdsorbedSolutionTheory.sp_res_pressure

AdsorbedSolutionTheory.x_sol

AdsorbedSolutionTheory.DualSiteLangmuir — Type

DualSiteLangmuir <: IsothermModel

DualSiteLangmuir(M1, K₀1, E1, M2, K₀2, E2)

DualSiteLangmuir(M1, K₀1, E1, M2, K₀2, E2) represents the DualSite Langmuir isotherm model.

Inputs

M1::T: maximum loading capacity of the first adsorbent site, [mol/kg]

K₀1::T: equilibrium constant at zero coverage for the first site, [1/Pa]

E1::T: adsorption energy for the first site, [J/mol]

M2::T: maximum loading capacity of the second adsorbent site, [mol/kg]

K₀2::T: equilibrium constant at zero coverage for the second site, [1/Pa]

E2::T: adsorption energy for the second site, [J/mol]

Description

The DualSite Langmuir equation is given by:

n = M1 * log(1 + K1 * p) + M2 * log(1 + K2 * p)

where:

n is the loading of the adsorbate on the adsorbent,

M1 and M2 are the maximum loading capacities of the first and second adsorbent sites, respectively,

K1 and K2 are the equilibrium constants at zero coverage for the first and second sites, respectively,

p is the pressure of the gas.

The adsorption energies E1 and E2 are related to the equilibrium constants K₀1 and K₀2 by the equations:

K1 = exp(-E1 / (R * T)) K2 = exp(-E2 / (R * T))

where:

R is the gas constant,

T is the temperature,

AdsorbedSolutionTheory.Langmuir — Type

`Langmuir(M, K₀, E)`

Langmuir <: IsothermModel

Langmuir(M, K₀, E) represents the Langmuir isotherm model, which describes the adsorption of a gas on a solid surface.

Inputs

M::T: maximum loading capacity of the adsorbent, [mol/kg]

K₀::T: equilibrium constant at zero coverage, [1/Pa]

E::T: adsorption energy, [J/mol]

Description

The Langmuir equation is given by:

n = (M * K₀ * p) / (1 + K₀ * p)

where:

n is the loading of the adsorbate on the adsorbent,

M is the maximum loading capacity of the adsorbent,

K₀ is the equilibrium constant at zero coverage,

p is the pressure of the gas.

The adsorption energy E is related to the equilibrium constant K₀ by the equation:

K₀ = exp(-E / (R * T))

where:

R is the gas constant,

T is the temperature.

AdsorbedSolutionTheory.Toth — Type

Toth <: IsothermModel

Toth(M, K₀, E, f₀, β)

Inputs

M::T: maximum loading capacity of the adsorbent, [mol/kg]

K₀::T: equilibrium constant at zero coverage,[1/Pa]`

E::T: adsorption energy, [J/mol]

f₀::T: Empirical parameter, -

β::T: Empirical parameter, K

Description

K = K₀exp(-E/(RT)) f = f₀ + β/T nᵢ = MKP/(1 + (KP)ᶠ)¹/ᶠ

AdsorbedSolutionTheory.Unilan — Type

`Unilan(M, K₀, E)`

Unilan <: IsothermModel

Unilan(M, K₀, E, f) represents the Unilan (UNIform distribution LANgmuir) isotherm model, which describes the adsorption of a gas on a solid surface.

Inputs

M::T: maximum loading capacity of the adsorbent, [mol/kg]

K₀::T: equilibrium constant at zero coverage, [1/Pa]

E::T: adsorption energy, [J/mol]

f::T: heterogeneity of the adsorbent (no units)

Description

The UNILAN equation is given by:

n = M * log((1 + K₀* exp(f) * p)/(1 + K₀ * exp(-f) * p)) / (2 * f)

where:

n is the loading of the adsorbate on the adsorbent,

M is the maximum loading capacity of the adsorbent,

K₀ is the equilibrium constant at zero coverage,

p is the pressure of the gas.

f is the heterogeneity of the adsorbent. at the limit f -> 0, the langmuir isotherm is recovered.

The adsorption energy E is related to the equilibrium constant K₀ by the equation:

K₀ = exp(-E / (R * T))

where:

R is the gas constant,

T is the temperature.

AdsorbedSolutionTheory.f∂f — Method

f∂f(f,x)

returns f and ∂f/∂x evaluated in x, using ForwardDiff.jl, DiffResults.jl and StaticArrays.jl to calculate everything in one pass.

AdsorbedSolutionTheory.f∂f∂2f — Method

f∂f∂2f(f,x)

returns f,∂f/∂x,and ∂²f/∂²x and evaluated in x, using ForwardDiff.jl, DiffResults.jl and StaticArrays.jl to calculate everything in one pass.

AdsorbedSolutionTheory.henry_coefficient — Method

henry_coefficient(model, T)

default units: [mol/kg]

Returns the single component spreading pressure of the model given the temperature T, defined as:

\[H = 2\]

AdsorbedSolutionTheory.isosteric_heat — Method

`isosteric_heat(model::IsothermModel, Vᵍ, Vᵃ = zero(eltype(model)), p, T)`

Calculate the isosteric heat of adsorption for a given isotherm model.

Inputs

model::IsothermModel: The isotherm model used to describe the adsorption process.

Vᵍ: The molar volume of the gas phase.

Vᵃ: The molar volume of the adsorbed phase (normaly Vᵃ << Vᵍ, default is zero).

p: Pressure at which the isosteric heat is evaluated.

T: Temperature at which the isosteric heat is evaluated.

Returns

The estimated isosteric heat of adsorption.

Description

The function Estimates the isosteric heat of adsorption for a single component from its isotherm using the Clausius-Clapeyron Equation:

Q_st = T × (Vᵍ - Vᵃ) × (∂n∂T)ₚ/(∂n∂P)ₜ (for explicit loading expressions)

Pan et al. (1998) https://doi.org/10.1021/la9803373

AdsorbedSolutionTheory.loading — Method

loading(model::IsothermModel, p, T)

Calculate the loading based on the model, pressure (p), and temperature (T).

Inputs

model::IsothermModel: the isotherm model

AdsorbedSolutionTheory.nlsolve — Function

function nlsolve(f!,x0,method=TrustRegion(Newton(), Dogleg()), options=NEqOptions(),chunk = ForwardDiff.Chunk{2}())

Given a function f!(result,x) that returns a system of equations, nlsolve(f!,x0) returns a NLSolvers.ConvergenceInfo struct that contains the results of the non-linear solving procedure.

Uses NLSolvers.jl as backend, the jacobian is calculated with ForwardDiff.jl, with the specified chunk size

To obtain the underlying solution vector, use x_sol

To see available solvers and options, check NLSolvers.jl

AdsorbedSolutionTheory.sp_res_pressure — Method

sprespressure(model::IsothermModel,q)

given an isotherm::IsothermModel and Π = spres(model,p), find p such that spres(model,p) = Π. by default, it performs a root-finding over the isotherm

AdsorbedSolutionTheory.x_sol — Method

x_sol(res::NLSolvers.ConvergenceInfo)

Returns the scalar or vector x that solves the system of equations or is the minimizer of an optimization procedure.