Reference

Contents

Reference

Contents

Index

Index

Langmuir.f∂f

Langmuir.f∂f∂2f

Langmuir.henry_coefficient

Langmuir.iast

Langmuir.isosteric_heat

Langmuir.isotherm_lower_bound

Langmuir.isotherm_upper_bound

Langmuir.loading

Langmuir.nlsolve

Langmuir.pressure

Langmuir.saturated_loading

Langmuir.sp_res

Langmuir.x_sol

Langmuir.@MultiSite

Langmuir.@with_metadata

Langmuir.nlsolve — Function

function nlsolve(f!,x0,method=TrustRegion(Newton(), Dogleg()), options=NEqOptions(),chunk = ForwardDiff.Chunk{2}())

Given a function f!(result,x) that returns a system of equations, nlsolve(f!,x0) returns a NLSolvers.ConvergenceInfo struct that contains the results of the non-linear solving procedure.

Uses NLSolvers.jl as backend, the jacobian is calculated with ForwardDiff.jl, with the specified chunk size

To obtain the underlying solution vector, use x_sol

To see available solvers and options, check NLSolvers.jl

Langmuir.@MultiSite — Macro

@MultiSite(isotherms)

Utility macro to build MultiSite types.

Example:

julia> v1 = @MultiSite{LangmuirS1,LangmuirS1} #abstract eltype
MultiSite{T, Tuple{LangmuirS1{T}, LangmuirS1{T}}} where T

julia> Langmuir.from_vec(v,[1,2,0,3,4,0])
MultiSite{Int64, Tuple{LangmuirS1{Int64}, LangmuirS1{Int64}}}((LangmuirS1{Int64}(1, 2, 0), LangmuirS1{Int64}(3, 4, 0)))

julia> v2 = @MultiSite{LangmuirS1, LangmuirS1}{Float64} #concrete eltype
MultiSite{Float64, Tuple{LangmuirS1{Float64}, LangmuirS1{Float64}}}

julia> Langmuir.from_vec(v2,[1,2,0,3,4,0])
MultiSite{Float64, Tuple{LangmuirS1{Float64}, LangmuirS1{Float64}}}((LangmuirS1{Float64}(1.0, 2.0, 0.0), LangmuirS1{Float64}(3.0, 4.0, 0.0)))

Langmuir.isotherm_lower_bound — Function

isotherm_lower_bound(model::IsothermModel)
isotherm_lower_bound(T,model::IsothermModel)
isotherm_lower_bound(T,::Type{M}) where M <:IsothermModel

Returns the lower bound for the parameters of the isotherm model model of type M. with number type T, as a Ntuple{model_length(M),T}. The default assumes that all parameters are nonnegative.

Langmuir.isosteric_heat — Function

isosteric_heat(model::IsothermModel, Vᵍ, p, T; Vᵃ = zero(eltype(model))) -> Qₛₜ

Calculate the isosteric heat of adsorption for a given isotherm model.

Arguments

model::IsothermModel: The isotherm model used to describe the adsorption process.

Vᵍ: The molar volume of the gas phase.

Vᵃ: The molar volume of the adsorbed phase (typically Vᵃ << Vᵍ; default is zero).

p: Pressure at which the isosteric heat is evaluated.

T: Temperature at which the isosteric heat is evaluated.

Returns

Qₛₜ: The estimated isosteric heat of adsorption.

Description

The function estimates the isosteric heat of adsorption Qₛₜ for a single component using its isotherm and the Clausius-Clapeyron equation:

Qₛₜ = -T * (Vᵍ - Vᵃ) * (∂n/∂T)ₚ / (∂n/∂p)ₜ

where:

n is the loading,

Vᵍ is the molar volume of the gas phase,

Vᵃ is the molar volume of the adsorbed phase,

T is the temperature,

p is the pressure.

This equation is derived based on the Clausius-Clapeyron relation, which relates the temperature dependence of the loading to the isosteric heat.

References:

Pan, H., Ritter, J. A., & Balbuena, P. B. (1998). Examination of the approximations used in determining the isosteric heat of adsorption from the Clausius−Clapeyron equation. Langmuir: The ACS Journal of Surfaces and Colloids, 14(21), 6323–6327. doi:10.1021/la9803373

Langmuir.f∂f∂2f — Function

f∂f∂2f(f,x)

returns f,∂f/∂x,and ∂²f/∂²x and evaluated in x, using ForwardDiff.jl, DiffResults.jl and StaticArrays.jl to calculate everything in one pass.

Langmuir.pressure — Function

pressure(model::IsothermModel, x, T, f; approx = :exact)

given an isotherm::IsothermModel and x = f(model,p,T), find p such that f(model,p,T) = x. There are two options for f and x:

when f is sp_res, then x = sp_res(model, p, T)

when f is loading, then x = loading(model, p, T)

By default, it performs a root-finding over the isotherm. but custom implementations can be done by overloading pressure_impl(model::MyModel,x,T,f::typeof(f),approx) The approx::Symbol argument indicates if the procedure is exact or approximate. by default a henry coefficient aproximation is used when approx =:henry is used.

Langmuir.saturated_loading — Function

saturated_loading(model::IsothermModel, T)

Returns the loading of of an adsorption isotherm when the pressure tends to infinity. by default it is evaluated at 1/√eps(eltype(model)) (6.7108864e7 for Float64 inputs.).

Inputs

model::IsothermModel: the isotherm model

T: temperature.

Langmuir.x_sol — Function

x_sol(res::NLSolvers.ConvergenceInfo)

Returns the scalar or vector x that solves the system of equations or is the minimizer of an optimization procedure.

Langmuir.loading — Function

loading(model::IsothermModel, p, T) -> q

Calculate the loading q based on the provided isotherm model, pressure p, and temperature T.

Arguments

model::IsothermModel: An instance of IsothermModel, representing the isotherm model to be used for the calculation.

p: The pressure at which the loading is to be calculated.

T: The temperature at which the loading is to be calculated.

Returns

q: The calculated loading based on the isotherm model, pressure, and temperature.

Description

This function computes the loading q based on the given isotherm model, pressure p, and temperature T.

Langmuir.henry_coefficient — Function

henry_coefficient(model::IsothermModel, T) -> H

Calculate the Henry's coefficient for a single component system using the specified isotherm model and temperature T.

Arguments

model::IsothermModel: An instance of IsothermModel, representing the isotherm model to be used for the calculation.

T: The temperature at which the Henry's coefficient is to be calculated.

Returns

H: The Henry's coefficient in the default units of [mol/kg].

Description

This function returns the Henry's coefficient, which is a measure of the initial slope of the adsorption isotherm at low pressures. It is defined as the derivative of the loading q with respect to pressure p at p = 0:

H = (∂q/∂p) at p = 0 at a given T.

Langmuir.isotherm_upper_bound — Function

isotherm_upper_bound(model::IsothermModel)
isotherm_upper_bound(T,model::IsothermModel)
isotherm_upper_bound(T,::Type{M}) where M <:IsothermModel

Returns the upper bound for the parameters of the isotherm model model of type M. with number type T, as a Ntuple{model_length(M),T}. The default assumes no upper bound for the parameters.

Langmuir.sp_res — Function

sp_res(model::IsothermModel, p, T) -> Π

Calculate the reduced spreading pressure for a given isotherm model at a specific pressure p and temperature T.

Arguments

model::IsothermModel: An instance of IsothermModel, representing the isotherm model used for the calculation.

p: The pressure at which the reduced spreading pressure is to be calculated.

T: The temperature at which the reduced spreading pressure is to be calculated.

Returns

Π: The reduced spreading pressure

Description

The reduced spreading pressure is a key quantity in Ideal Adsorbed Solution Theory (IAST), used to describe the adsorption behavior of mixtures. This function calculates the reduced spreading pressure Π by integrating the isotherm equation over the pressure range from 0 to p.

The reduced spreading pressure is often calculated numerically as:

Π = ∫ (q(p') / p') dp' from 0 to p

where:

q(p') is the loading at pressure p'.

Langmuir.iast — Function

iast(models,p,T,y,method = FastIAS(),gas_model = nothing;x0 = nothing,maxiters = 100,reltol = 1e-12, abstol = 1e-10)

TODO: docs

returns qtot,x,convergencesymbol (:success, or :maxiters_exceeded)

Langmuir.@with_metadata — Macro

@with_metadata(struct_expr)

macro that allows to define an isotherm model with additional metadata, about parameter bounds and descriptions of parameters:

Usage:

Langmuir.@with_metadata struct MyIsotherm{T} <: IsothermModel{T}
    A::T,(0,1),"field A" #bounds and description provided
    B::T #nothing provided
    C::T,(1,10) #only bounds provided
    D::T,nothing,"field D" #only description provided
end

from_vec(MyIsotherm,(1,2,3,4)) #ok
from_vec(MyIsotherm,(-1,2,3,4)) #ArgumentError: MyIsotherm: value for the field `A` (field A) is out of the parameter bounds: (0.0 <= -1 <= 1.0) == false
from_vec(MyIsotherm,(1,2,-3,4)) #ArgumentError: MyIsotherm: value for the field `C` is out of the parameter bounds: (1.0 <= -3 <= 10.0) == false

Langmuir.f∂f — Function

f∂f(f,x)

returns f and ∂f/∂x evaluated in x, using ForwardDiff.jl, DiffResults.jl and StaticArrays.jl to calculate everything in one pass.