Supported Isotherms

Here you can find the list of supported isotherm models.

Supported Isotherms

No-parameter Isotherm Models

One-parameter Isotherm Models

Three-parameter Isotherm Models

Four or more parameter Isotherm Models

Langmuir.BrunauerEmmettTeller

Langmuir.Freundlich

Langmuir.Henry

Langmuir.LangmuirFreundlich

Langmuir.LangmuirS1

Langmuir.MultiSite

Langmuir.Quadratic

Langmuir.RedlichPeterson

Langmuir.Sips

Langmuir.ThermodynamicLangmuir

Langmuir.Toth

Langmuir.Unilan

Langmuir.ZeroIsotherm

No-parameter Isotherm Models

Langmuir.ZeroIsotherm — Type

`ZeroIsotherm()`

ZeroIsotherm <: IsothermModel

ZeroIsotherm represents an isotherm that has no adsorption. it can be used to define carrier gases.

Description

A zero isotherm is defined by:

n = 0

where:

n is the loading of the adsorbate on the adsorbent

One-parameter Isotherm Models

Langmuir.Henry — Type

`Henry(Kh, E)`

Henry <: IsothermModel

Inputs

Kh::T: Affinity parameter, or Henry’s constant at T → ∞ [mol/kg/Pa]

E::T: Adsorption energy, [J/mol]

Description

The Henry isotherm model describes the adsorption of gases on solid surfaces at low pressures, where the amount of adsorbate adsorbed is directly proportional to the pressure of the gas. This model is typically valid in the low-pressure limit of adsorption, where the adsorption sites are far from saturation and the interactions between adsorbed molecules are negligible.

The adsorption behavior follows Henry's law:

n = K * p

The energy parameter E is related to Henry's constant Kh by the equation:

K = Kh*exp(-E / (R * T))

Where:

R is the universal gas constant, [J/mol/K],

T is the temperature, [K].

Three-parameter Isotherm Models

Langmuir.LangmuirS1 — Type

`LangmuirS1(M, K₀, E)`

LangmuirS1 <: IsothermModel

LangmuirS1(M, K₀, E) represents the single site Langmuir isotherm model.

Inputs

M::T: Saturation loading, [mol/kg]

K₀::T: Affinity parameter at T → ∞, [1/Pa]

E::T: Adsorption energy, [J/mol]

Description

The LangmuirS1 equation is given by:

n = (M * K * p) / (1 + K * p)

The adsorption energy E is related to the equilibrium constant K₀ by the equation:

K = K₀*exp(-E / (R * T))

where:

R is the gas constant,

T is the temperature.

Four or more parameter Isotherm Models

Langmuir.Freundlich — Type

`Freundlich(K₀, f₀, β, E)`

 Freundlich <: IsothermModel

Inputs

K₀::T: Affinity parameter at T → ∞, [1/Pa]

f₀::T: Surface heterogeneity parameter at T → ∞, [-]

β::T: Surface heterogeneity coefficient [K]

E::T: Adsorption energy, [J/mol]

Description

The Freundlich isotherm is given by:

n = K_f × pᶠ

The affinity parameter K_f is a temperature dependent and can be linked to adsorption energy E by:

K_f = K₀ × exp(-E / (RT))

The exponent f is also temperature dependent and can be expressed as:

f = f₀ - β/T

Where:

R is the gas constant,

T is the temperature.

Langmuir.RedlichPeterson — Type

`RedlichPeterson(M, K₀, E, f₀, β)`

RedlichPeterson <: IsothermModel

Inputs

M::T: Saturation loading, [mol/kg]

K₀::T: Affinity parameter at T → ∞, [Pa]¹/ᶠ

E::T: Adsorption energy, [J/mol]

f₀::T: Surface heterogeneity parameter at T → ∞, [-]

β::T: Surface heterogeneity coefficient, [K]

Description

The RedlichPeterson equation is given by:

n = M * p / (1 + K₀ * p^f)

The adsorption energy E is related to the equilibrium constant K₀ by the equation:

K = K₀ × exp(-E / (R * T))

The exponent f is also temperature dependent and can be expressed as:

f = f₀ - β/T

where:

R is the gas constant,

T is the temperature.

Langmuir.Toth — Type

`Toth(M, K₀, E, f₀, β)`

Toth <: IsothermModel

Inputs

M::T: Saturation loading, [mol/kg]

K₀::T: Affinity parameter at T → ∞, [Pa⁻¹]

E::T: Adsorption energy, [J/mol]

f₀::T: Surface heterogeneity parameter at T → ∞, [-]

β::T: Surface heterogeneity coefficient, [K]

Description

Toth isotherm model:

n = MKP/(1 + (K*P)ᶠ)¹/ᶠ

The adsorption energy E is related to the equilibrium constant K₀ by the equation:

K = K₀ × exp(-E / (R * T))

The exponent f is also temperature dependent and can be expressed as:

f = f₀ - β/T

where:

R is the gas constant,

T is the temperature.

Langmuir.Unilan — Type

`Unilan(M, K₀, E, f₀, β)`

Unilan <: IsothermModel

Inputs

M::T: Saturation loading, [mol/kg]

K₀::T: Affinity parameter at T → ∞, [Pa⁻⁻¹]`

E::T: Adsorption energy, [J/mol]

f₀::T: Surface heterogeneity parameter at T → ∞, [-]

β::T: Surface heterogeneity coefficient, [K]

Description

The UNILAN equation is given by:

n = M * log((1 + K* exp(f) * p)/(1 + K * exp(-f) * p)) / (2 * f)

The adsorption energy E is related to the equilibrium constant K₀ by the equation:

K = K₀ × exp(-E / (R * T))

The surface heterogeneity parameter f is also temperature-dependent and can be expressed as:

f = f₀ - β / T

where:

R is the gas constant,

T is the temperature.

Langmuir.Quadratic — Type

`Quadratic(K₀a, K₀b, M, Ea, Eb)`

Quadratic <: IsothermModel

Inputs

K₀a::T: Affinity parameter A at T → ∞, [1/Pa]

K₀b::T: Affinity parameter B at T → ∞, [1/Pa^2]

M::T: Saturation loading, [mol/kg]

Ea::T: Adsorption energy A, [J/mol]

Eb::T: Adsorption energy B, [J/mol]

Description

The Quadratic isotherm model is given by:

n = M × (Ka + 2Kb × p) × p / (1 + p × (Ka + Kb × p))

The model assumes that the affinity parameters Ka and Kb are temperature-dependent and follow the relation:

Ka = K₀a * exp(-Ea / (RT))

Kb = K₀b * exp(-Eb / (RT))

Where:

Ka and Kb are the affinity parameters at temperature T,

R is the gas constant,

T is the absolute temperature.

Langmuir.BrunauerEmmettTeller — Type

`BrunauerEmmettTeller(K₀a, K₀b, M, E)`

BrunauerEmmettTeller <: IsothermModel

The Brunauer-Emmett-Teller (BET) isotherm model describes multilayer adsorption on a homogeneous surface. It is widely used to characterize adsorption behavior in mesoporous materials.

Inputs

K₀a::T: Affinity parameter A at T → ∞, [1/Pa]

K₀b::T: Affinity parameter A at T → ∞, [1/Pa]

M::T: Saturation loading, [mol/kg]

E::T: Adsorption energy, [J/mol]

Description

The BET equation is given by:

n = M × K₀a × p / (1 - K₀b × p) × (1 + (K₀a - K₀b) × p)

Temperature dependence:

The affinity parameters K₀a and K₀b depend on temperature and are given by the following expressions:

K₀a = K₀a × exp(-E / (RT))

K₀b = K₀b × exp(-E / (RT))

Where:

R is the gas constant,

T is the absolute temperature.

Langmuir.Sips — Type

`Sips(M, K₀, E, f₀, β)`

Sips <: IsothermModel

Inputs

M::T: Saturation loading, [mol/kg]

K₀::T: Affinity parameter at T → ∞, [Pa⁻¹]

E::T: Adsorption energy, [J/mol]

f₀::T: Surface heterogeneity parameter at T → ∞, [-]

β::T: Surface heterogeneity coefficient, [K]

Description

The Sips equation is given by:

n = M * (K * p)ᶠ / (1 + (K * p)ᶠ)

The adsorption energy E is related to the equilibrium constant K₀ by the equation:

K = K₀ × exp(-E / (R * T))

The exponent f is also temperature dependent and can be expressed as:

f = f₀ - β/T

where:

R is the gas constant,

T is the temperature.

Langmuir.LangmuirFreundlich — Type

`LangmuirFreundlich(M, K₀, E, f₀, β)`

LangmuirFreundlich <: IsothermModel

Inputs

M::T: Saturation loading, [mol/kg]

K₀::T: Affinity parameter at T → ∞, [1/Pa]

E::T: Adsorption energy, [J/mol]

f₀::T: Surface heterogeneity parameter at T → ∞, [-]

β::T: Surface heterogeneity coefficient, [K]

Description

The Langmuir-Freundlich form of the isotherm is:

n = M × K × pᶠ /(1 + K × pᶠ)

Where:

n is the loading of the adsorbate on the adsorbent,

Temperature dependence:

The affinity parameter K is temperature-dependent and can be expressed as:

K = K₀ × exp(-E / (RT))

The surface heterogeneity parameter f is also temperature-dependent and can be expressed as:

f = f₀ - β / T

Where:

R is the gas constant,

T is the absolute temperature.

Langmuir.MultiSite — Type

MultiSite(isotherms::Vararg{IsothermModel}...)

given a list of isotherms, create a multisite isotherm model.

model1 = LangmuirS1(3.0,1.0,0.0)
model2 = LangmuirS1(3.0,0.9,3000.0)
double_site_LangmuirS1 = MultiSite(model1,model2) #create a multisite model with two LangmuirS1 isotherms

Langmuir.ThermodynamicLangmuir — Type

ThermodynamicLangmuir(M, K₀, E, Bᵢᵩ)

The ThermodynamicLangmuir struct represents a thermodynamic Langmuir model with activity coefficients calculated using an aNRTL-like approach.

Parameters

M: Saturation loading [mol/kg]

K₀: Affinity parameter at high temperature [1/Pa]

E: Adsorption energy [J/mol]

Bᵢᵩ: Interaction energy parameter between adsorbate species i and vacant sites [J/mol]

Description

The Langmuir equation is given by:

nᵢ = (M * K * P) / (γᵢ/γᵩ + K * P)

where nᵢ is the adsorbed amount of species i, K is calculated as:

K = K₀ * exp(-E / (R * T))

where:

R is the gas constant

T is the temperature

The activity coefficients γᵢ and γᵩ are determined using the Gibbs excess free energy, gᴱ/RT, which is calculated based on the surface fractions (θᵢ, θᵩ) and interaction parameters derived from Bᵢᵩ. This free energy value is used in the activity_coefficient function to compute the activity coefficients of the adsorbate and phantom molecules.